requirements.txt
google-api-core==2.16.2
google-auth==2.27.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.40.0
google-cloud-automl==2.13.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.0
google-cloud-storage==2.14.0
pandas==2.0.3
